requirements.txt
flask>=3.1.2
requests>=2.0.0
dnspython>=2.4.2
python-whois>=0.7.3
flask-limiter>=3.7.0
beautifulsoup4>=4.12.0
